+++ satkit/helpers/processing_helpers.py
@@ -1,8 +1,8 @@
 #
-# Copyright (c) 2019-2023 
+# Copyright (c) 2019-2024
 # Pertti Palo, Scott Moisik, Matthew Faytak, and Motoki Saito.
 #
-# This file is part of Speech Articulation ToolKIT 
+# This file is part of Speech Articulation ToolKIT
 # (see https://github.com/giuthas/satkit/).
 #
 # This program is free software: you can redistribute it and/or modify
@@ -39,7 +39,7 @@ def product_dict(**kwargs):
 
     ```python
     options = {"number": [1,2,3], "color": ["orange","blue"] }
-    print(list(product_dict(options)))
+    print(list(product_dict(**options)))
 
     [ {"number": 1, "color": "orange"},
     {"number": 1, "color": "blue"},