+++ docs/Release_notes.markdown
@@ -1,5 +1,34 @@
 # SATKIT Release Notes
 
+## Foreshadowing the next release: Version 0.10.0
+
+- New functionality and configuration for extracting annotations and producing
+  plots out of them.
+- New two stage internal configuration model
+  - Started rolling out a model where configuration files are still parsed with
+    strictyaml to preserve comments in round tripping (which has not yet been
+    implemented), but get reinterpreted as pydantic classes to make coding
+    easier.
+
+Known issues:
+
+- To be determined
+
+## Version 0.9.0
+
+- Simulated data and sensitivity analysis for metrics
+  - Two contours for running sensitivity simulations for contour metrics.
+  - Perturbation generation for the contours.
+  - Functions for running metrics on the simulated data.
+  - Lots of plotting routines to look at the results.
+  
+Known issues:
+
+Same as in 0.8.0.
+
+- Some perturbation related plotting functions have hard-coded subplot
+  divisions because Comparison is not yet sortable.
+
 ## Version 0.8.0
 
 - Splines